requirements.txt
Flask==2.2.5
Flask_SQLAlchemy==3.0.3
Werkzeug==2.2.3
requests==2.31.0
